+++ src/compas/colors/colormap.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 
 import os
 
-from compas.utilities import linspace
+from compas.itertools import linspace
 
 from .color import Color
 from .mpl_colormap import _magma_data