+++ src/compas/geometry/shapes/cone.py
@@ -7,7 +7,7 @@ from math import pi
 from math import sin
 from math import sqrt
 
-from compas.utilities import pairwise
+from compas.itertools import pairwise
 from compas.geometry import transform_points
 from compas.geometry import Circle
 from compas.geometry import Plane